Fix clickEvent.js linters errors
